Use size_t instead of int in genx

+++ xml/details/genx/char-props.c
@@ -18,16 +18,16 @@ static void charProp(char * p, int c, int prop)
   p[c] |= prop;
 }
 
-static void rangeProp(char * p, int start, int end, int prop)
+static void rangeProp(char * p, size_t start, size_t end, int prop)
 {
-  int i;
+  size_t i;
   for (i = start; i <= end; i++)
     p[i] |= prop;
 }
 
 void genxSetCharProps(char * p)
 {
-  int i;
+  size_t i;
 
   for (i = 0; i < GENX_CHAR_TABLE_SIZE; i++)
     p[i] = 0;
@@ -392,4 +392,3 @@ void genxSetCharProps(char * p)
 
 #endif /* GENX_CHAR_TABLE_SIZE == 0x10000 */
 }
-
